forensik-panel/versions/v1/requirements.txt
Flask==3.0.3
Jinja2==3.1.4
Werkzeug==3.0.4
pymssql==2.3.2
polars==1.5.0
duckdb==1.1.0
xlsxwriter==3.2.0
PyYAML==6.0.2
python-dotenv==1.0.1
waitress==3.0.0
tabulate==0.9.0
